+++ flashfold/scripts/__init__.py
@@ -3,4 +3,7 @@ from .extend_main_sequence_db import *
 from .download_sequence_db import *
 from .predict_structure import *
 from .download_ncbi_assembly import *
+from .create_json_with_ligand import *
+from .run_alphafold3 import *
+from .predict_stoi import *
 from .generate_report import *