+++ flashfold/tools/__init__.py
@@ -1,1 +1,3 @@
 from .workflow import *
+from .util import *
+from .modify_json import *