+++ flashfold/tools/workflow.py
@@ -1,18 +1,22 @@
 import os
-from flashfold.utils import run_jobs_in_parallel, run_single_job, get_sequence_length_from_single_fasta
+from typing import List, Dict, Optional
+from flashfold.utils import run_jobs_in_parallel, run_single_job, get_sequence_length_from_single_fasta, \
+    update_time_log
 
 
 min_jackhmmer_hits = 1000
 
 
-def run_homology_search(fasta_files: list, database_fasta: str, provided_cpu: int, out_path: str) -> None:
+def run_homology_search(fasta_files: List, database_fasta: str, provided_cpu: int, compact_msa: bool,
+                        out_path: str) -> None:
     """
     Run homology searching.
 
     Args:
-        fasta_files (list): List of paths to FASTA files.
+        fasta_files (List): List of paths to FASTA files.
         database_fasta (str): Path to the database FASTA file.
         provided_cpu (int): Number of CPU cores to use.
+        compact_msa (bool): Flag to indicate if compact MSA is to be generated.
         out_path (str): Output directory path.
 
     Returns:
@@ -23,6 +27,7 @@ def run_homology_search(fasta_files: list, database_fasta: str, provided_cpu: in
     deduplicate_script = os.path.join(current_file_dir, "deduplicate_msa.py")
     combined_commands = []
     cpu_per_job = min(8, provided_cpu)
+    dedup_compact_msa_flag = "" if compact_msa is False else "--compact"
     for fasta_file in fasta_files:
         fasta_seq_length = get_sequence_length_from_single_fasta(fasta_file)
         fasta_basename = os.path.basename(fasta_file)
@@ -38,18 +43,18 @@ def run_homology_search(fasta_files: list, database_fasta: str, provided_cpu: in
                        "-o /dev/null --cpu %s -A %s %s %s" % (cpu_per_job, sto_file_path, fasta_file, database_fasta))
         reformat_sto_command = ("python3 %s %s --output_a3m %s --output_fas %s"
                                 % (reformat_script, sto_file_path, a3m_file_path, fas_file_path))
-        deduplicate_command = ("python3 %s %s --query_length %s --min_sequences %s --output %s --threads %s"
+        deduplicate_command = ("python3 %s %s --query_length %s --min_sequences %s --output %s --threads %s %s"
                                % (deduplicate_script, fas_file_path, fasta_seq_length, min_jackhmmer_hits,
-                                   diverse_hits_file_path, cpu_per_job))
+                                   diverse_hits_file_path, cpu_per_job, dedup_compact_msa_flag))
         combined_command = f"{sto_command} && {reformat_sto_command} && {deduplicate_command}"
         combined_commands.append(combined_command)
     run_jobs_in_parallel(provided_cpu, cpu_per_job, combined_commands, "Homology searching")
     return
 
 
-def run_af3tools(a3m_file_path: str, json_file_path: str) -> None:
+def run_msa_to_json(a3m_file_path: str, json_file_path: str) -> None:
     """
-    Run af3tools for creating json file from msa for AF3 input.
+    Run msa_to_json.py for creating json file from msa for AF3 input.
 
     Args:
         a3m_file_path (str): Path to the A3M file.
@@ -59,9 +64,60 @@ def run_af3tools(a3m_file_path: str, json_file_path: str) -> None:
         None
     """
     current_file_dir = os.path.dirname(os.path.abspath(__file__))
-    af3tools_script = os.path.join(current_file_dir, "msa_to_json.py")
-    af3tools_command = f"python3 {af3tools_script} -i {a3m_file_path} -o {json_file_path}"
-    run_single_job(af3tools_command, "JSON file creation")
+    msa_to_json_script = os.path.join(current_file_dir, "msa_to_json.py")
+    msa_to_json_script_command = f"python3 {msa_to_json_script} -i {a3m_file_path} -o {json_file_path}"
+    run_single_job(msa_to_json_script_command, "JSON file creation")
     return
 
 
+def run_alphafold3_docker(config: Dict[str, str], query_files: List[str], output_dir: str,
+                          log_file_path: Optional[str]) -> None:
+    """
+    Run AlphaFold3 using Docker.
+
+    Args:
+        config (Dict[str, str]): Configuration dictionary containing paths to AlphaFold3 parameters, database,
+        and Docker image.
+        query_files (List[str]): List of paths to query JSON files.
+        output_dir (str): Directory to store the output results.
+        log_file_path (Optional[str]): Path to the log file for recording timings.
+
+    Returns:
+        None
+    """
+
+    af3_params_path = config.get('params')
+    af3_db_path = config.get('db')
+    af3_image_name = config.get('image')
+
+    for query_file in query_files:
+        query_file_basename = os.path.basename(query_file)
+        query_file_without_ext = os.path.splitext(query_file_basename)[0]
+        query_file_dir = os.path.dirname(query_file)
+        af3_docker_command = [
+            "docker", "run", "-it",
+            "--user", f"{os.getuid()}:{os.getgid()}",
+            "--volume", f"{query_file_dir}:/input",
+            "--volume", f"{output_dir}:/output",
+            "--volume", f"{af3_params_path}:/models",
+            "--volume", f"{af3_db_path}:/public_databases",
+            "--gpus", "all",
+            f"{af3_image_name}",
+            "python", "run_alphafold.py",
+            f"--json_path=/input/{query_file_basename}",
+            "--model_dir=/models",
+            "--output_dir=/output",
+            "--run_data_pipeline=False"
+        ]
+        af3_docker_command_str = " ".join(af3_docker_command)
+
+        if log_file_path:
+            update_time_log(log_file_path, f"Started: Structure prediction for {query_file_without_ext}",
+                            True)
+        run_single_job(af3_docker_command_str, f"Running AlphaFold3 for {query_file_without_ext}")
+
+        if log_file_path:
+            update_time_log(log_file_path, f"Completed: Structure prediction for {query_file_without_ext}",
+                            True)
+
+    return